allow persisting local backup files

+++ services/backer/dump.sh
@@ -11,6 +11,10 @@ pg_dump $DB_URL -f "$FILE"
 
 curl -F file=@$FILE $BACKUPPER_URL/backups/$SERVICE_NAME/$FILE\?token\=$TOKEN_SECRET
 
-rm $FILE
+
+if [ $REMOVE_LOCAL_FILE_AFTER_BACKUP = "true" ]
+then
+  rm $FILE
+fi
 
 echo "Job finished: $(date)"
